Modified TestWrapper.cpp

+++ PowerRangerMain/EmptyGeneralTesting/SPA/VarTable.cpp
@@ -26,7 +26,7 @@ INDEX VarTable::insertVar(VARNAME varName) {
 // Returns the name of a variable at VarTable [ind]
 // If ‘ind’ is out of range, error (or throw exception)
 VARNAME VarTable::getVarName (INDEX ind){
-	if (ind >= varTable.size()) {
+	if (ind >= (signed int) varTable.size()) {
 		// throw error
 	}
 	return varTable[ind];
